requirements.txt
pandas
plotly
requests
streamlit
google-cloud-bigquery
pandas-gbq
tqdm
openpyxl